Feat: mypage 라우팅 구현중
- account.js 생성
- mypage.js의 계좌관리 부분 account.js로 이동
